Extract createFeatureSlice from projectStore (store-slice template)
